Moved register/signin to Authentication Cleaned up imports in app.py
