+++ routes/app.py
@@ -1,25 +1,19 @@
-from flask import Flask, render_template, redirect, request, url_for, session
+from flask import Flask, render_template, redirect, request, session
 import os.path
-import pandas 
 import pandas as pd
-import csv
-
 import os, sys
+from model.Authentication.Register import validatePasswordStrength, emailValidator, ensurePasswordsAreEqual, registerNewUser, checkIfEmailExists
+from model.Authentication.SignIn import verifyEmailAndPassword, checkEmailExists, signInUser
+from model.Movie import MovieFactory
+from model.Admin.AddMovie import addMovieToCSV
+
+
 currDir = os.path.dirname(os.path.realpath(__file__))
 rootDir = os.path.abspath(os.path.join(currDir, '..'))
 if rootDir not in sys.path: # add parent dir to paths
     sys.path.append(rootDir)
 
-from model.Register import validatePasswordStrength, emailValidator, ensurePasswordsAreEqual,\
-    registerNewUser, checkIfEmailExists
-
-from model.SignIn import verifyEmailAndPassword, checkEmailExists, signInUser
-
-from model.Movie import MovieFactory
-
 movieFactory = MovieFactory.MovieFactory()
-
-from model.Admin.AddMovie import addNewReleaseMovie, addChildrensMovie, addStandardMovie, addMovieToCSV
 
 TEMPLATES_PATH_STRING = str(os.path.abspath('..')) + '/templates'
 STATIC_PATH_STRING = str(os.path.abspath('..')) + '/static'
